Add -Wshorten-64-to-32 to core/fpdfdoc.
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/89590
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfdoc/cpdf_bookmark.cpp
@@ -30,12 +30,12 @@ WideString CPDF_Bookmark::GetTitle() const {
     return WideString();
 
   WideString title = pString->GetUnicodeText();
-  int len = title.GetLength();
+  size_t len = title.GetLength();
   if (!len)
     return WideString();
 
   std::vector<wchar_t, FxAllocAllocator<wchar_t>> buf(len);
-  for (int i = 0; i < len; i++) {
+  for (size_t i = 0; i < len; i++) {
     wchar_t w = title[i];
     buf[i] = w > 0x20 ? w : 0x20;
   }
